+++ dart/collision/DistanceResult.cpp
@@ -29,12 +29,53 @@
  *   POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef DART_COLLISION_RESULT_HPP_
-#define DART_COLLISION_RESULT_HPP_
+#include "dart/collision/DistanceResult.hpp"
 
-#warning "This header has been deprecated in DART 6.1. "\
-  "Please include CollisionResult.hpp intead."
+#define DART_DEFAULT_MIN_DISTANCE (0.0)
+#define DART_DEFAULT_UNCLAMPED_MIN_DISTANCE (0.0)
 
-#include "dart/collision/CollisionResult.hpp"
+namespace dart {
+namespace collision {
 
-#endif  // DART_COLLISION_RESULT_HPP_
+//==============================================================================
+DistanceResult::DistanceResult()
+  : minDistance(DART_DEFAULT_MIN_DISTANCE),
+    unclampedMinDistance(0.0),
+    shapeFrame1(nullptr),
+    shapeFrame2(nullptr),
+    nearestPoint1(Eigen::Vector3d::Zero()),
+    nearestPoint2(Eigen::Vector3d::Zero())
+{
+  // Do nothing
+}
+
+//==============================================================================
+void DistanceResult::clear()
+{
+  minDistance = DART_DEFAULT_MIN_DISTANCE;
+  unclampedMinDistance = DART_DEFAULT_UNCLAMPED_MIN_DISTANCE;
+
+  nearestPoint1.setZero();
+  nearestPoint2.setZero();
+
+  shapeFrame1 = nullptr;
+  shapeFrame2 = nullptr;
+}
+
+//==============================================================================
+bool DistanceResult::found() const
+{
+  if (!shapeFrame1 || !shapeFrame2)
+    return false;
+
+  return true;
+}
+
+//==============================================================================
+bool DistanceResult::isMinDistanceClamped() const
+{
+  return found() && (minDistance == unclampedMinDistance);
+}
+
+}  // namespace collision
+}  // namespace dart