+++ dart/dynamics/BoxShape.cpp
@@ -36,7 +36,7 @@ namespace dart {
 namespace dynamics {
 
 BoxShape::BoxShape(const Eigen::Vector3d& _size)
-  : Shape(BOX),
+  : Shape(),
     mSize(_size) {
   assert(_size[0] > 0.0);
   assert(_size[1] > 0.0);